refactor: separate Focus Group and Roving Tabindex into class extensions
